Ursa Ult Implemented

+++ Legacy of the Rift/Assets/Scripts/Enemies/Enemy.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health = 100;
+    public float health = 100;
 
     //public GameObject deathEffect;
     public Transform player;
@@ -18,7 +18,7 @@ public class Enemy : MonoBehaviour
         player = GameObject.FindGameObjectWithTag("Player").GetComponent<Transform>();
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
         StartCoroutine(Slowed());
